+++ trunk/linbox/linbox/randiter/modular-crooked.h
@@ -5,20 +5,20 @@
  * Adapted from randiter/modular-dense.h
  * by Brice Boyer <brice.boyer@imag.fr>
  *
- * 
+ *
  * ========LICENCE========
  * This file is part of the library LinBox.
- * 
+ *
  * LinBox is free software: you can redistribute it and/or modify
  * it under the terms of the  GNU Lesser General Public
  * License as published by the Free Software Foundation; either
  * version 2.1 of the License, or (at your option) any later version.
- * 
+ *
  * This library is distributed in the hope that it will be useful,
  * but WITHOUT ANY WARRANTY; without even the implied warranty of
  * MERCHANTABILITY or FITNESS FOR A PARTICULAR PURPOSE.  See the GNU
  * Lesser General Public License for more details.
- * 
+ *
  * You should have received a copy of the GNU Lesser General Public
  * License along with this library; if not, write to the Free Software
  * Foundation, Inc., 51 Franklin Street, Fifth Floor, Boston, MA  02110-1301  USA
@@ -94,7 +94,7 @@ namespace LinBox
 			<< " and seed " << _seed << std::endl;
 
 			// Seed random number generator
-			srand (_seed);
+			srand ((unsigned int)_seed);
 		}
 
 		/** Copy constructor.
@@ -134,7 +134,7 @@ namespace LinBox
 		 * @return reference to random field element
 		 */
 		Element &random (Element &a) const
-		{ return _field.init(a,rand()); }
+		{ return _field.init(a,(double)rand()); }
 
 		/** Random field element creator.
 		 * This returns a random field element from the information supplied
